add an optional robot_prefix to the provider configuration (#494)
add an optional robot_prefix to the provider configuration

`robot_prefix` (String) Without this option, the provider will try to
automatically determine the robot prefix with a call to the admin api.
If you don't have admin access and want to create system robot account,
you'll have to set this value.

and fix, permission access effect, to be able to use the provider with
robot account

Signed-off-by: flbla <flbla@users.noreply.github.com>