Allow login via OIDC session ID cookie (sid)
This is a workaround for various auth issues with the Harbor API when using OIDC without having access to the OIDC client secret and ID or to a local user account.

It also adds the FAQ-recommended retry to handle CSRF.

For example, in corporate environments you might have a shared harbor instance for which you want to be able to automate project creation etc., but don't have access to the OIDC client directly; in my particular case we have a GitLab auth and the secret and token exchange happens between GitLab and Harbor, the latter providing a logged in user only with a session (cookie).

The following non-exhaustive list of issues are related:

- https://github.com/goharbor/terraform-provider-harbor/issues/383 Adds the ability to use bearer_tokens, which is not possible without access to it
- https://github.com/goharbor/harbor/issues/16398 Not being able to get a Bearer token for robots and not being able to use Basic auth due to OIDC login
- https://github.com/goharbor/harbor/issues/16647 Only able to access some API routes via token, but usually via sid

The FAQ also recommend to use the bearer token:
https://github.com/goharbor/harbor/wiki/Harbor-FAQs#api

The FAQ also notes that
> XSRF protection is enabled in Harbor V1.10. After this version, when you
> issue an API call, you can [repeat with the CSRF tokens]

This commit adds exactly that:

- It allows to authenticate via a session ID cookie (which must be retrieved out of band, as handling OIDC is out of scope of this commit)
- It handles CSRF-Token failures by retrying with extracted tokens or erroring out if no token is present
- Adds documentation
- Adds tests for the csrf token extraction

I tested it with our GitLab + Harbor setup and the following resources and their operations:

- Adding and removing projects
- Adding and removing project members
- Adding and removing robot accounts

The easiest way to get an sid= cookie is to login via your favorite browser and copy the value from your cookies.

To automate it (which is the point of using terraform), you can export your sid as the HARBOR_SESSION_ID, for example by doing some curl calls (here an example with GitLab and "selecting" the relevant login option):

    HARBOR_LOGIN_URL=https://yourharbor.example.com/c/oidc/login\?redirect_url\=/harbor/projects
    GITLAB_SIGNIN_URL=https://yourgitlab.example.com/users/sign_in
    FHG_GITLAB_USERNAME=gitlab username
    FHG_GITLAB_PASSWORD=gitlab password

    authenticity_token=$(curl --cookie-jar cookiejar --silent --location ${HARBOR_LOGIN_URL} | grep /users/sign_in | grep authenticity_token | cut -d" " -f 11 | cut -d'"' -f 2)
    redirect_url=$(curl --cookie cookiejar --cookie-jar cookiejar --silent --location ${GITLAB_SIGNIN_URL} --form "authenticity_token=${authenticity_token}" --form "user[login]=${FHG_GITLAB_USERNAME}" --form "user[password]=${FHG_GITLAB_PASSWORD}" --form 'user[remember_me]="0"' | grep redirectUri | head -n 1 | cut -d'"' -f 2)
    export HARBOR_SESSION_ID=$(curl --cookie cookiejar --cookie-jar cookiejar --silent --include $redirect_url | grep "Set-Cookie: sid" | cut -d";" -f 1 | cut -d"=" -f 2)

Signed-off-by: Sebastian Höffner <info@sebastian-hoeffner.de>