fix(ConfigAuth): Drop LDAP default params from oidcConflictsWith (#520)
This drops the following params from the `oidcConflictsWith` list:
- `ldap_verify_cert`
- `ldap_scope`
- `ldap_group_membership`
- `ldap_group_scope`

In the ConfigAuth schema, some LDAP params are defined as both optional
and with default values. This has a knock-on effect during schema
validation where the presence of these default values for params
returned by `oidcConflictsWith()` to have a conflict under the hood.

In the underlying API client, LDAP and OIDC params are marshalled and
segmented by auth type for both get:
https://github.com/goharbor/terraform-provider-harbor/blob/main/client/config.go#L46-L90
and set calls:
https://github.com/goharbor/terraform-provider-harbor/blob/main/client/config.go#L123-L162

Whether these params should be truly optional or have default values is
a separate question that this PR doesn't intend to address but instead
focuses on resolving the more immediate ConflictsWith issue.

Fixes #461

Signed-off-by: Paul Bailey <spacepants@users.noreply.github.com>